+++ CHANGELOG.md
@@ -15,6 +15,6 @@ Support for Rails 3.x. For 2.x support, check out the branch `rails_2_3` or vers
 
 #### Fixed
 
- * The `has_and_belongs_to_many` associations, for secondbase models.
+ * The `has_and_belongs_to_many` associations, for SecondBase models.
  * Patched `ActiveRecord::TestFixtures` so that transactional fixture support is respected for the SecondBase.
  * Reorganized monkey patches to make it easier to work in fixes for different versions of Rails.